user-service/requirements.txt
asyncpg==0.20.1
fastapi==0.48.0
uvicorn==0.11.2
firebase-admin==6.0.1
pyjwt==2.6.0
bcrypt==3.2.2
pydantic==1.10.2